// taurus-protect-sdk-typescript/src/internal/openapi/apis/WebhookCallsApi.ts
export interface WebhookServiceGetWebhookCallsRequest {
    eventID?: string;
    webhookID?: string;
    status?: string;
    cursorCurrentPage?: string;
    cursorPageRequest?: string;
    cursorPageSize?: string;
    sortOrder?: string;
}

/**
 * WebhookCallsApi - interface
 * 
 * @export
 * @interface WebhookCallsApiInterface
 */
export interface WebhookCallsApiInterface {
    /**
     * This endpoint lists calls that have been made to webhooks. Required Role: **Admin**, **AdminReadOnly.
     * @summary Get webhook calls
     * @param {string} [eventID] 
     * @param {string} [webhookID] 
     * @param {string} [status] 
     * @param {string} [cursorCurrentPage] Base64-encoded string representing the current window of data
     * @param {string} [cursorPageRequest] The page to request, w.r.t the current page. Can be one of &#x60;FIRST&#x60;, &#x60;PREVIOUS&#x60;, &#x60;NEXT&#x60;, &#x60;LAST&#x60;
     * @param {string} [cursorPageSize] The size of the page requested. The handling service should impose a hard limit on this
     * @param {string} [sortOrder] Webhook calls are sorted by creation date and ID. Set this to ASC for an ascending order or DESC for descending. The default value is DESC
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof WebhookCallsApiInterface
     */
    webhookServiceGetWebhookCallsRaw(requestParameters: WebhookServiceGetWebhookCallsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetWebhookCallsReply>>;

    /**
     * This endpoint lists calls that have been made to webhooks. Required Role: **Admin**, **AdminReadOnly.
     * Get webhook calls
     */
    webhookServiceGetWebhookCalls(requestParameters: WebhookServiceGetWebhookCallsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetWebhookCallsReply>;

}

/**
 * 
 */
export class WebhookCallsApi extends runtime.BaseAPI implements WebhookCallsApiInterface {

    /**
     * This endpoint lists calls that have been made to webhooks. Required Role: **Admin**, **AdminReadOnly.
     * Get webhook calls
     */
    async webhookServiceGetWebhookCallsRaw(requestParameters: WebhookServiceGetWebhookCallsRequest, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<runtime.ApiResponse<TgvalidatordGetWebhookCallsReply>> {
        const queryParameters: any = {};

        if (requestParameters['eventID'] != null) {
            queryParameters['eventID'] = requestParameters['eventID'];
        }

        if (requestParameters['webhookID'] != null) {
            queryParameters['webhookID'] = requestParameters['webhookID'];
        }

        if (requestParameters['status'] != null) {
            queryParameters['status'] = requestParameters['status'];
        }

        if (requestParameters['cursorCurrentPage'] != null) {
            queryParameters['cursor.currentPage'] = requestParameters['cursorCurrentPage'];
        }

        if (requestParameters['cursorPageRequest'] != null) {
            queryParameters['cursor.pageRequest'] = requestParameters['cursorPageRequest'];
        }

        if (requestParameters['cursorPageSize'] != null) {
            queryParameters['cursor.pageSize'] = requestParameters['cursorPageSize'];
        }

        if (requestParameters['sortOrder'] != null) {
            queryParameters['sortOrder'] = requestParameters['sortOrder'];
        }

        const headerParameters: runtime.HTTPHeaders = {};

        if (this.configuration && this.configuration.apiKey) {
            headerParameters["Authorization"] = await this.configuration.apiKey("Authorization"); // ApiKeyTPV1 authentication
        }

        const response = await this.request({
            path: `/api/rest/v1/webhooks/calls`,
            method: 'GET',
            headers: headerParameters,
            query: queryParameters,
        }, initOverrides);

        return new runtime.JSONApiResponse(response, (jsonValue) => TgvalidatordGetWebhookCallsReplyFromJSON(jsonValue));
    }

    /**
     * This endpoint lists calls that have been made to webhooks. Required Role: **Admin**, **AdminReadOnly.
     * Get webhook calls
     */
    async webhookServiceGetWebhookCalls(requestParameters: WebhookServiceGetWebhookCallsRequest = {}, initOverrides?: RequestInit | runtime.InitOverrideFunction): Promise<TgvalidatordGetWebhookCallsReply> {
        const response = await this.webhookServiceGetWebhookCallsRaw(requestParameters, initOverrides);
        return await response.value();
    }

}